# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Kanban API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/lists": {
            "get": {
                "tags": [
                    "Lists"
                ],
                "summary": "Get list of all board's lists",
                "description": "Returns list of all board's lists",
                "operationId": "f977b65239f52b76b1aab3ccbf78b663",
                "responses": {
                    "200": {
                        "description": "Array[] : contain all lists"
                    },
                    "419": {
                        "description": "Delay error : CSRF Token missed ?"
                    }
                }
            },
            "post": {
                "tags": [
                    "Lists"
                ],
                "summary": "Insert new list",
                "description": "Create new list to DB",
                "operationId": "f711288a7f01987eebe161c229f110cb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "String -> modified list id"
                    },
                    "400": {
                        "description": "Bad request : No name || not connected"
                    },
                    "419": {
                        "description": "Delay error : CSRF Token missed ?"
                    }
                }
            }
        },
        "/api/lists/{id}": {
            "get": {
                "tags": [
                    "Lists"
                ],
                "summary": "Get info for specific list by id",
                "description": "Returns list by id",
                "operationId": "6d080e643faf7717d8c908600dc8927d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Array[] : Target list"
                    },
                    "419": {
                        "description": "Delay error : CSRF Token missed ?"
                    }
                }
            },
            "put": {
                "tags": [
                    "Lists"
                ],
                "summary": "Modify info for specific list by id",
                "description": "Modify list by id",
                "operationId": "12cbd776b9a2c84d7d2e72eb736e6e87",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of target list",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "is_archived": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "String -> modified list id"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "419": {
                        "description": "Delay error : CSRF Token missed ?"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Lists"
                ],
                "summary": "Delete list",
                "description": "Delete list by id",
                "operationId": "1eee913566619f184990ff756abac89a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of target list",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Method not ready now"
                    }
                }
            }
        },
...
    }
}